+++ apps/ced-portal-be/src/application/use-cases/opportunities/get-operator-opportunity.use-case.ts
@@ -35,10 +35,9 @@ export const makeGetOperatorOpportunityUseCase =
   async (input) =>
     validateUseCaseInput(GetOperatorOpportunityInputSchema, input).andThen(
       (validatedInput) =>
-        new ResultAsync(opportunityRepository.getById(validatedInput)).andThen(
-          (data) =>
-            data
-              ? ok(data)
-              : err(new NotFoundError("Opportunity", "not found")),
+        new ResultAsync(
+          opportunityRepository.findByIdAndOperatorId(validatedInput),
+        ).andThen((data) =>
+          data ? ok(data) : err(new NotFoundError("Opportunity", "not found")),
         ),
     );