+++ packages/io-core-adapter-tracing/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pagopa/io-core-adapter-tracing
 
+## 0.0.3
+
+### Patch Changes
+
+- d146483: integrated error logging and audits
+
+## 0.0.2
+
+### Patch Changes
+
+- c92a9f0: Abstracted service name into infra
+
 ## 0.0.1
 
 ### Patch Changes