+++ packages/io-core-domain/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pagopa/io-core-domain
 
+## 0.0.4
+
+### Patch Changes
+
+- 40aaab7: Added string utils
+
 ## 0.0.3
 
 ### Patch Changes